Add Pickaxe Breaking
Adding the ability for a pickaxe to break, increasing the cooldown duration. It was necessary for the cooldowns dictionary values to be converted from an integer to a dictionary that stores both the timestamp of the last roll and the pickaxe state.

Reordered the cooldown and the mining frenzy check code to ensure a single point for the cooldown to be calculated from.